Create package Duke
