+++ src/main/java/duke/Deadline.java
@@ -1,48 +1,52 @@
+package duke;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
+import java.time.format.DateTimeFormatterBuilder;
+import java.util.Date;
 
 /**
- * This is a child class of Task, Event.
- * Event accepts another variable, 'at' that
- * stores the time this Event is taking place
+ * This is a child class of duke.Task, duke.Deadline.
+ * duke.Deadline accepts another variable, 'by' that
+ * stores the time this duke.Deadline has to be completed
  *
  * @author  Hsiao Jiet
  * @version 1.0
  * @since   2022-1-15
  */
 
-public class Event extends Task {
-    protected LocalDate at;
+public class Deadline extends Task {
+    protected LocalDate by;
 
     /**
-     * Creates Event object with provided LocalDate variable
+     * Creates duke.Deadline object with provided LocalDate variable
      */
-    public Event(String n, int d, LocalDate a) {
+    public Deadline(String n, int d, LocalDate b) {
         super(n, d);
-        super.type = 'E';
-        at = a;
+        super.type = 'D';
+        by = b;
     }
 
     /**
-     * Creates Event object without provided LocalDate variable
+     * Creates duke.Deadline object without provided LocalDate variable
      * Used when there is the need to parse text into date first
      */
-    public Event(String n, int d) {
+    public Deadline(String n, int d) {
         super(n, d);
-        super.type = 'E';
+        super.type = 'D';
     }
 
-    public void setAt(LocalDate at) {
-        this.at = at;
+    public void setBy(LocalDate by) {
+        this.by = by;
     }
 
     /**
-     * Setting Event object's date it is held on
+     * Setting duke.Deadline object's date it should end by
      */
     public void setStringToLocalDate(String date) {
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy/MMMM/dd");
         LocalDate localDate = LocalDate.parse(date, formatter);
-        setAt(localDate);
+        setBy(localDate);
     }
 
     /**
@@ -50,19 +54,20 @@ public class Event extends Task {
      */
     public String convertLocalDateToString() {
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy/MMMM/dd");
-        return at.format(formatter);
+        return by.format(formatter);
     }
 
     /**
-     * Producing a user-friendly view of an Event task
+     * Producing a user-friendly view of a duke.Deadline task
      */
     @Override
     public String toString() {
         StringBuilder res = new StringBuilder();
         res.append(getTaskIcon()).append(" - ");
         res.append(this.getDoneIcon()).append(" - ");
-        res.append(this.name).append(" - ");;
+        res.append(this.name).append(" - ");
         res.append(convertLocalDateToString()).append("\n");
         return res.toString();
     }
+
 }
