+++ src/main/java/duke/Event.java
@@ -1,50 +1,50 @@
+package duke;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeFormatterBuilder;
-import java.util.Date;
 
 /**
- * This is a child class of Task, Deadline.
- * Deadline accepts another variable, 'by' that
- * stores the time this Deadline has to be completed
+ * This is a child class of duke.Task, duke.Event.
+ * duke.Event accepts another variable, 'at' that
+ * stores the time this duke.Event is taking place
  *
  * @author  Hsiao Jiet
  * @version 1.0
  * @since   2022-1-15
  */
 
-public class Deadline extends Task {
-    protected LocalDate by;
+public class Event extends Task {
+    protected LocalDate at;
 
     /**
-     * Creates Deadline object with provided LocalDate variable
+     * Creates duke.Event object with provided LocalDate variable
      */
-    public Deadline(String n, int d, LocalDate b) {
+    public Event(String n, int d, LocalDate a) {
         super(n, d);
-        super.type = 'D';
-        by = b;
+        super.type = 'E';
+        at = a;
     }
 
     /**
-     * Creates Deadline object without provided LocalDate variable
+     * Creates duke.Event object without provided LocalDate variable
      * Used when there is the need to parse text into date first
      */
-    public Deadline(String n, int d) {
+    public Event(String n, int d) {
         super(n, d);
-        super.type = 'D';
+        super.type = 'E';
     }
 
-    public void setBy(LocalDate by) {
-        this.by = by;
+    public void setAt(LocalDate at) {
+        this.at = at;
     }
 
     /**
-     * Setting Deadline object's date it should end by
+     * Setting duke.Event object's date it is held on
      */
     public void setStringToLocalDate(String date) {
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy/MMMM/dd");
         LocalDate localDate = LocalDate.parse(date, formatter);
-        setBy(localDate);
+        setAt(localDate);
     }
 
     /**
@@ -52,20 +52,19 @@ public class Deadline extends Task {
      */
     public String convertLocalDateToString() {
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy/MMMM/dd");
-        return by.format(formatter);
+        return at.format(formatter);
     }
 
     /**
-     * Producing a user-friendly view of a Deadline task
+     * Producing a user-friendly view of an duke.Event task
      */
     @Override
     public String toString() {
         StringBuilder res = new StringBuilder();
         res.append(getTaskIcon()).append(" - ");
         res.append(this.getDoneIcon()).append(" - ");
-        res.append(this.name).append(" - ");
+        res.append(this.name).append(" - ");;
         res.append(convertLocalDateToString()).append("\n");
         return res.toString();
     }
-
 }
